My take on Presenters. See full commit message.
I've forked Florian and Niko's code to serve our needs at Fridge Door,
Inc. The code as of this commit basically works, but shows little in the
way of optimization and refinement. I have not updated the tests because
I am not a good person and do not use tests to drive development. I've
so far updated the inline documentation, but not the README.

Most of the interesting features I've added have to do with better
supporting presenter hierarchies and inheritance. As with the original
code, I am principally concerned with wrapping individual model objects
and homogeneous collections of model objects.

- If a presenter is not found for a model's class, we traverse up its
  inheritance until we find a presenter for some superclass.
- Presenter subclasses append a view directory to ActionView's search
  path, allowing them to override superclasses' templates.
- Render calls are dispatched by method_missing and look like
  render_partial_name. This allows presenters to alter or reroute a
  render call, e.g. rendering a partial of a different name, or
  dispatching to a different presenter instance.
- Presenter mixins may also provide their own template directories to be
  included in the search path.

This is definitely a work in progress, but I welcome any feedback about
the approach, and especially about making interfacing sanely with
ActionView and ActionController.